Move Brain and DNA into the creature plugin
