+++ src/modules/plugins/creatures/components/Brain.js
