+++ src/modules/plugins/creatures/components/DNA.js
