+++ src/modules/plugins/creatures/systems/touchProcessor.test.js
@@ -1,5 +1,5 @@
 import TouchProcessor from "./TouchProcessor";
-import Brain from "../../core/components/Brain";
+import Brain from "../../creatures/components/Brain";
 import HexGrid from "../../../grid/HexGrid";
 import Coord from "../../core/components/Coord";
 import { expect } from "chai";
